Esercizio completato

+++ assets/js/script.js
@@ -7,6 +7,8 @@ var erroreElenco;
 var elenco = [];
 var newNome;
 var newCognome;
+var toggle = true;
+var modificaId;
 
 window.addEventListener("DOMContentLoaded", init);
 
@@ -23,7 +25,11 @@ function init() {
 
 function eventHandler() {
 	addBtn.addEventListener("click", function () {
-		controlla();
+		if (toggle) {
+			controlla();
+		} else {
+			modifica(modificaId);
+		}
 	});
 }
 
@@ -41,7 +47,7 @@ function printData() {
 					elencoHTML.innerHTML += `
 					<li class="list-group-item">
 
-					<button type="button" class="edit btn btn-primary fs-5" onClick="openEdit(${element.id})"><i class="bi bi-pencil-square"></i></button>
+					<button type="button" class="edit btn btn-primary fs-5" onClick="openEdit(${element.id}, '${element.nome}', '${element.cognome}')"><i class="bi bi-pencil-square"></i></button>
 
 					<button type="button" class="delete btn btn-danger me-4 fs-6" onClick="elimina(${element.id})"><i class="bi bi-x-lg"></i></button>
 
@@ -73,7 +79,16 @@ function controlla() {
 		};
 		addData(data);
 	} else {
-		errore.innerHTML = `<i class="bi bi-exclamation-triangle-fill"></i> Attenzione! Compilare correttamente i campi! <i class="bi bi-exclamation-triangle-fill"></i>`;
+
+		// Avviso
+		errore.innerHTML = `<i class="bi bi-exclamation-triangle-fill"></i> Attenzione! Compilare correttamente i campi! <i class="bi bi-exclamation-triangle-fill"></i>`;;
+
+		// Bordo rosso input
+		document.getElementById("nome").classList.remove("border-primary", "border-opacity-25");
+		document.getElementById("nome").classList.add("border-danger", "border-opacity-75");
+		document.getElementById("cognome").classList.remove("border-primary", "border-opacity-25");
+		document.getElementById("cognome").classList.add("border-danger", "border-opacity-75");
+
 		return;
 	}
 }
@@ -95,63 +110,83 @@ function clearForm() {
 }
 
 
-// DELETE
 
+// DELETE
 function elimina(data) {
+
+	if (!window.confirm("Sei sicur*?")) {
+		return;
+	};
+
 	fetch('http://localhost:3000/elenco/' + data, {
 		method: 'DELETE'
 	});
+
 	clearForm();
 }
 
 // PUT
 function modifica(data) {
-	newNome = document.getElementById('newNome');
-	newCognome = document.getElementById('newCognome');
 
-	if (newNome.value != '' && newCognome.value != '') {
-		fetch('http://localhost:3000/elenco/' + data, {
-			method: 'PUT',
-			headers: {
-				'Content-Type': 'application/json;charset=utf-8',
-			},
-			body: JSON.stringify({ nome: `${newNome.value}`, cognome: `${newCognome.value}` })
-		});
-
-		clearForm();
-	} else {
-		// Bordo rosso input
-		document.getElementById("newNome").classList.remove("border-primary", "border-opacity-25");
-		document.getElementById("newNome").classList.add("border-danger", "border-opacity-75");
-		document.getElementById("newCognome").classList.remove("border-primary", "border-opacity-25");
-		document.getElementById("newCognome").classList.add("border-danger", "border-opacity-75");
+	if (nome.value == '' || cognome.value == '') {
 
 		// Avviso
-		document.getElementById("attenzione").innerHTML = `<i class="bi bi-exclamation-triangle-fill"></i> Attenzione! Compilare correttamente i campi! <i class="bi bi-exclamation-triangle-fill"></i>`;
+		errore.innerHTML = `<i class="bi bi-exclamation-triangle-fill"></i> Attenzione! Compilare correttamente i campi! <i class="bi bi-exclamation-triangle-fill"></i>`;;
+
+		// Bordo rosso input
+		document.getElementById("nome").classList.remove("border-primary", "border-opacity-25");
+		document.getElementById("nome").classList.add("border-danger", "border-opacity-75");
+		document.getElementById("cognome").classList.remove("border-primary", "border-opacity-25");
+		document.getElementById("cognome").classList.add("border-danger", "border-opacity-75");
+
 		return;
-	}
+	};
+
+	if (!window.confirm("Modificare l'utente?") && !toggle) {
+		return;
+	};
+
+
+
+	fetch('http://localhost:3000/elenco/' + data, {
+		method: 'PUT',
+		headers: {
+			'Content-Type': 'application/json;charset=utf-8',
+		},
+		body: JSON.stringify({ nome: `${nome.value}`, cognome: `${cognome.value}` })
+	});
+
+	clearForm();
+
+	toggle = true;
 }
 
-function openEdit(data) {
-	// Fai apparire il div di modifica
-	// if (document.getElementById("divModifica").style.display == "block") {
-	// 	document.getElementById("divModifica").style.display = "none";
-	// } else {
-	// 	document.getElementById("divModifica").style.display = "block";
-	// }
+function openEdit(dataId, dataName, dataSurname) {
 
-	document.getElementById("divModifica").style.display = "block";
+	// Metti i values dentro gli input
+
+	toggle = false;
+
+	modificaId = dataId;
+
+	nome.value = dataName;
+	cognome.value = dataSurname;
 
 	// Metti questo button nel sotto-div di modifica
 
-	if (!document.getElementById("modifica")) {
-		document.getElementById("subModifica").innerHTML += `
-		<button type="button" class="btn btn-outline-primary" id="modifica" onclick = "modifica(${data})">MODIFICA</button>
-		`;
-
+	if (toggle == false) {
 		// Evidenzia ciò che si sta modificando
-		document.querySelector(`li:nth-of-type(${data})`).style.opacity = "0.5";
-		document.querySelector(`li:nth-of-type(${data})`).style.fontStyle = "italic";
-		document.querySelector(`li:nth-of-type(${data})`).style.border = "2px dashed #0D6EFD";
+		document.querySelector(`li:nth-of-type(${dataId})`).style.opacity = "0.5";
+		document.querySelector(`li:nth-of-type(${dataId})`).style.fontStyle = "italic";
+		document.querySelector(`li:nth-of-type(${dataId})`).style.border = "2px dashed #0D6EFD";
+
+		// Cambia il colore del button
+		addBtn.classList.remove("btn-primary");
+		addBtn.classList.add("btn-info");
+		addBtn.innerHTML = "MODIFICA";
+
+		// Cambia il colore del bordo del fieldset
+		document.querySelector("fieldset").classList.add("border-info");
 	}
+
 }
